ITERATION 3: RECIPE PAGE CONTENT
Completed the first recipe page on the website for a delicious
and quick chicken and rice dish. All text is sourced from
https://www.campbells.com/recipes/15-minute-chicken-rice-dinner/

Added an image of the completed dish to the page. Added an h3 element
title descripton and below copied a short description of the dish into
a <p> element.

-Added an ingredients header with an unordered lists of needed ingredients

-Added a steps header with an unordered list of the steps needed to
actually prepare the dish